+++ backend/canisters/registry/CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Expose both heap and stable memory in metrics ([#5718](https://github.com/open-chat-labs/open-chat/pull/5718))
-- Calculate and expose circulating supply ([#5771](https://github.com/open-chat-labs/open-chat/pull/5771))
+- Calculate and expose total supply and circulating supply ([#5771](https://github.com/open-chat-labs/open-chat/pull/5771))
 
 ## [[2.0.1143](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1143-registry)] - 2024-04-19